Refactor, remove circular import error
