+++ api/util.py
@@ -1,7 +1,6 @@
 import datetime
 import enum as baseEnum
 import hashlib
-import hmac
 import json
 import mimetypes
 import os
@@ -9,18 +8,16 @@ import random
 import re
 import string
 import time
-import urlparse
 import uuid
-from urllib import quote, urlencode
 
 import requests
 
 import bson
+import config
 import fs.errors
 import fs.path
 import pymongo
 
-from . import config
 from .web import errors
 
 BYTE_RANGE_RE = re.compile(r'^(?P<first>\d+)-(?P<last>\d+)?$')
@@ -524,36 +521,3 @@ def add_container_type(request, result):
     """Adds a 'container_type' property to result if fw_container_type is set in the request environment."""
     if 'fw_container_type' in request.environ and isinstance(result, dict):
         result['container_type'] = request.environ['fw_container_type']
-
-
-def generate_signed_url(url, method='GET', expires_in=3600):
-    expires = int(time.time() + datetime.timedelta(seconds=expires_in).total_seconds())
-    url_parts = list(urlparse.urlparse(url))
-    url_parts[2] = quote(url_parts[2])
-    signature = hmac.new(
-        str(config.get_config()['core']['signed_url_secret']),
-        msg='{}:{}:{}:{}'.format(method, url_parts[1].rsplit(':', 1)[0], url_parts[2], expires),
-        digestmod=hashlib.sha256
-    ).hexdigest()
-    query = dict(urlparse.parse_qsl(url_parts[4]))
-    query['expires'] = expires
-    query['signature'] = signature
-    url_parts[4] = urlencode(query)
-    return urlparse.urlunparse(url_parts)
-
-def verify_signed_url(url, method):
-    url_parts = list(urlparse.urlparse(url))
-    query = dict(urlparse.parse_qsl(url_parts[4]))
-    expires = query.pop('expires')
-    signature = query.pop('signature')
-    url_parts[4] = urlencode(query)
-    url = urlparse.urlunparse(url_parts)
-    calc_signature = hmac.new(
-        str(config.get_config()['core']['signed_url_secret']),
-        msg='{}:{}:{}:{}'.format(method, url_parts[1].rsplit(':', 1)[0], url_parts[2], expires),
-        digestmod=hashlib.sha256
-    ).hexdigest()
-    if signature != calc_signature:
-        raise errors.APIPermissionException('Invalid siganture')
-
-    return True
